backend/djangoprojects/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
djoser==2.2.2
hashids==1.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
dj-database-url==2.1.0
django-cloudinary-storage==0.3.0
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==22.0.0
uvicorn==0.30.0
whitenoise==6.6.0
dj-static==0.0.6